third_party/WebKit/LayoutTests/svg/hixie/error/011-expected.txt

CONSOLE ERROR: line 1: Error: <svg> attribute viewBox: Expected number, "0 0,, 1 1".
